Move searchpath from header.c into the runtime
Merge the path-searching code from stdlib/header.c with the runtime's
version.
